requirements.txt
langchain-google-genai
langchain
streamlit
python-dotenv
PyMuPDF
duckduckgo-search
chromadb 
langchain-community
